Add files via upload
Reading and analyzing NASA'S GPM data from 2000 to 2021. Each daily data is contained in a single file with resolution of 0.1*0.1 degree. For the study months, 122 files are per year. And a total of 122*4 files are read using pipeline method. Then averaged for study locations using mask and result is plotted.